Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 
-int addition(int first, int second)
+float addition(int first, int second)
 {
     int counter= first + second;
     return counter;
 }
 
-int subtraction(int first, int second)
+float subtraction(int first, int second)
 {
     int counter= first - second;
     return counter;
 }
 
-int multiplication(int first, int second)
+float multiplication(int first, int second)
 {
     int counter= first * second;
     return counter;
 }
 
-int division(int first, int second)
+float division(int first, int second)
 {
     int counter= first / second;
     return counter;
@@ -29,23 +29,23 @@ int main(void)
     int first, second;
     char special_character;
     printf("");
-    scanf("%i %i %c", &first, &second, &special_character);
+    scanf("%f %f %c", &first, &second, &special_character);
 
     if (special_character== '+')
     {
-        printf("%i", addition(first, second));
+        printf("%f", addition(first, second));
     }
     else if (special_character== '-')
     {
-        printf("%i", subtraction(first, second));
+        printf("%f", subtraction(first, second));
     }
     else if (special_character== '*')
     {
-        printf("%i", multiplication(first, second));
+        printf("%f", multiplication(first, second));
     }
     else
     {
-        printf("%i", division(first, second));
+        printf("%f", division(first, second));
     }
 
     return 0;
